+++ src/related.js
@@ -1,9 +1,13 @@
 // @related [test](/src/related.test.js)
 
-/* Accepts a single string and returns an array of related file objects with `name` and `path` keys. */
-function related(text) {
+/*
+ * Accepts a single string and returns an array of related file objects with `name` and `path` keys.
+ * An optional config object can be passed as the second argument. If it contains a `projectRoot` key, then
+ * its value will be prepended each found path.
+ */
+function related(text, config) {
   return findAnnotatedLines(text).reduce((lines, line) => {
-    return lines.concat(findLinks(line));
+    return lines.concat(findLinks(line, config));
   }, []);
 }
 
@@ -17,13 +21,13 @@ function findAnnotatedLines(text) {
   }
 }
 
-function findLinks(line) {
+function findLinks(line, config) {
   const re = /[^\[]*\[([^\]]+)\]\(([^\)]+)\)/g;
   let result = [];
   let match;
 
   while ((match = re.exec(line))) {
-    result.push({ name: match[1], path: match[2] });
+    result.push({ name: match[1], path: applyProjectRoot(match[2], config) });
   }
 
   return result;
@@ -33,8 +37,18 @@ function isNonEmptyString(s) {
   return s && typeof s == "string" && !/^\s*$/.test(s);
 }
 
+function applyProjectRoot(path, config) {
+  if (path && path.startsWith("/")) {
+    const projectRoot = (config && config.projectRoot) || "";
+    return projectRoot.concat(path);
+  } else {
+    return path;
+  }
+}
+
 module.exports = {
   related,
+  _applyProjectRoot: applyProjectRoot,
   _findAnnotatedLines: findAnnotatedLines,
-  _findLinks: findLinks,
+  _findLinks: findLinks
 };