+++ src/related.test.js
@@ -1,4 +1,4 @@
-const { related, _findAnnotatedLines, _findLinks } = require("./related");
+const { related, _applyProjectRoot, _findAnnotatedLines, _findLinks } = require("./related");
 
 describe("related", () => {
   test("returns [] when the input is not a string", () => {
@@ -11,7 +11,7 @@ describe("related", () => {
   });
 
   test("returns [] when the input contains no annotations", () => {
-    expect(related("foo\nbar\n")).toStrictEqual([]);
+    expect(related("ant\nbat\n")).toStrictEqual([]);
   });
 
   test("simple annotation", () => {
@@ -25,7 +25,7 @@ describe("related", () => {
 
   test("annotation with extra text", () => {
     expect(
-      related("foo @related bar [test](/src/related.test.js) baz")
+      related("ant @related bat [test](/src/related.test.js) cat")
     ).toStrictEqual([
       {
         name: "test",
@@ -36,7 +36,7 @@ describe("related", () => {
 
   test("annotation with spaces in link name", () => {
     expect(
-      related("foo @related bar [unit test](/src/related.test.js) baz")
+      related("ant @related bat [unit test](/src/related.test.js) cat")
     ).toStrictEqual([
       {
         name: "unit test",
@@ -48,7 +48,7 @@ describe("related", () => {
   test("multiple links on the same line", () => {
     expect(
       related(
-        "foo @related bar [test](/src/related.test.js) glorp [css](/assets/related.css) baz"
+        "ant @related bat [test](/src/related.test.js) cat [css](/assets/related.css) dog"
       )
     ).toStrictEqual([
       {
@@ -66,11 +66,11 @@ describe("related", () => {
     expect(
       related(
         `
-          blah
-          foo @related bar [test](/src/related.test.js) glorp [css](/assets/related.css) baz
-          blah
-          @related fez [icon](/assets/related.png) quux
-          blah
+          ant
+          bat @related cat [test](/src/related.test.js) dog [css](/assets/related.css) eel
+          fox
+          @related gnu [icon](/assets/related.png) hen
+          imp
         `
       )
     ).toStrictEqual([
@@ -102,13 +102,13 @@ describe("findAnnotatedLines", () => {
   });
 
   test("returns [] when there are no annotated lines", () => {
-    expect(_findAnnotatedLines("foo\nbar\nbaz\n")).toStrictEqual([]);
+    expect(_findAnnotatedLines("ant\nbat\ncat\n")).toStrictEqual([]);
   });
 
   test("returns annotated lines", () => {
     expect(
-      _findAnnotatedLines("foo\n@related bar\nbaz @related baz\nfez\n")
-    ).toStrictEqual(["@related bar", "baz @related baz"]);
+      _findAnnotatedLines("ant\n@related bat\ncat @related dog\neel\n")
+    ).toStrictEqual(["@related bat", "cat @related dog"]);
   });
 });
 
@@ -118,21 +118,48 @@ describe("findLinks", () => {
   });
 
   test("returns the link when there is one", () => {
-    expect(_findLinks("[foo](/foo.txt)")).toStrictEqual([
-      { name: "foo", path: "/foo.txt" },
+    expect(_findLinks("[ant](/ant.txt)")).toStrictEqual([
+      { name: "ant", path: "/ant.txt" },
     ]);
 
-    expect(_findLinks("blah [foo](/foo.txt) blah")).toStrictEqual([
-      { name: "foo", path: "/foo.txt" },
+    expect(_findLinks("ant [bat](/bat.txt) cat")).toStrictEqual([
+      { name: "bat", path: "/bat.txt" },
     ]);
   });
 
   test("returns all the links", () => {
     expect(
-      _findLinks("blah [foo](/foo.txt) blah [bar](/bar.txt) blah")
+      _findLinks("ant [bat](/bat.txt) cat [dog](/dog.txt) eel")
     ).toStrictEqual([
-      { name: "foo", path: "/foo.txt" },
-      { name: "bar", path: "/bar.txt" },
+      { name: "bat", path: "/bat.txt" },
+      { name: "dog", path: "/dog.txt" },
     ]);
+  });
+
+  test("prepends the project root to paths that start with '/' when specified in '.config/related-files.txt'", () => {
+    expect(
+      _findLinks(
+        "ant [bat](/bat.txt) cat [dog](dog.txt) eel [fox](/fox/fox.txt) gnu",
+        { projectRoot: "/proj/root" }
+      )
+    ).toStrictEqual([
+      { name: "bat", path: "/proj/root/bat.txt" },
+      { name: "dog", path: "dog.txt" },
+      { name: "fox", path: "/proj/root/fox/fox.txt" },
+    ]);
+  });
+});
+
+describe("applyProjectRoot", () => {
+  test("returns the input when the config is undefined", () => {
+    expect(_applyProjectRoot("/ant/bat", undefined)).toStrictEqual("/ant/bat")
+  });
+
+  test("returns the input when the config does not have a 'project_root' key", () => {
+    expect(_applyProjectRoot("/ant/bat", {x: "y"})).toStrictEqual("/ant/bat")
+  });
+
+  test("prepends the project root", () => {
+    expect(_applyProjectRoot("/ant/bat", {projectRoot: "/cat/dog"})).toStrictEqual("/cat/dog/ant/bat")
   });
 });