+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## Version 1.0
+## Version 1.1.0
+
+- The `related` function now accepts a second argument, `config`. The only currently supported config value is
+  `projectRoot`, which if specified is prepended to each found path that starts with a "/".
+
+## Version 1.0.0
 
 Initial release.